pyproject.toml
[tool.poetry]
name = "api-life-is-strange"
version = "0.1.0"
description = ""
authors = ["daniwells <daniellimafreita@gmail.com>"]
license = "mit"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["standard"], version = "^0.115.2"}
uvicorn = "^0.32.0"
alembic = "^1.13.3"
sqlalchemy = "^2.0.36"
psycopg2 = "^2.9.10"
pytest = "^8.3.3"
python-multipart = "^0.0.12"
python-jose = "^3.3.0"
passlib = "^1.7.4"
pydantic = "^2.9.2"
python-decouple = "^3.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
